Add files via upload
This repository contains the source code for a landing page for an ecommerce website that sells a new laptop. The landing page is designed to be responsive and visually appealing, and it features high-quality image. The page also includes brief product information. The landing page is built using HTML, CSS